Moved all components tests into components folder and renamed them to remove prefix component*
