+++ test/component/viewpointTest.js
@@ -1,4 +1,4 @@
-let d3X3dom = require("../");
+let d3X3dom = require("../../build/d3-x3dom");
 let tape = require("tape");
 
 tape("Text viewpoint quickView, viewpoint.quickView(), ", function(test) {
